1) Improve the way I get output after the Lock/Unlock calls
Change TLockState to TLockInfo and add a new Enum: TLockState which defines: (Locked, Unlocked, INITIAL) to TLockInfo in order
to better control the output from both, the lock and unlock calls. Also update the PrintLockInfo procedure to also handle the now the case with the TLockState enum. Add aswell a procedure
named WriteEmptyLines to write as many empty new lines as needed. Update also the IndentifyLockInfo and IndentifyUnlockInfo to handle the TLockState enum.